% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchfor.R
\name{searchfor}
\alias{searchfor}
\title{Search for pattern}
\usage{
searchfor(text, pattern = NULL)
}
\arguments{
\item{text}{corpus}

\item{pattern}{search for}
}
\value{
string
}
\description{
Search for pattern
}

